backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose
PyJWT
